remove abstract-milvus-vectordb.ts to avoid issue in vscode
