+++ packages/core/src/vectordb/index.ts
@@ -7,12 +7,6 @@ export {
     COLLECTION_LIMIT_MESSAGE
 } from './types';
 
-// Re-export abstract base class and config
-export {
-    AbstractMilvusVectorDatabase,
-    BaseMilvusConfig
-} from './abstract-milvus-vectordb';
-
 // Implementation class exports
 export { MilvusRestfulVectorDatabase, MilvusRestfulConfig } from './milvus-restful-vectordb';
 export { MilvusVectorDatabase, MilvusConfig } from './milvus-vectordb';
